refactor: types 폴더 분리
