+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useCardNumber';
 export * from './useCardExpiration';
 export * from './useCardCVC';
-export * from './types';
